[rename]`SimpleProductDto`위치 변경 (#41)
`SimpleProductDto`의 위치를 `dto`패키지로 이동
