+++ src/main/java/org/kakaoshare/backend/domain/product/repository/query/ProductRepositoryCustom.java
@@ -2,7 +2,7 @@ package org.kakaoshare.backend.domain.product.repository.query;
 
 import org.kakaoshare.backend.domain.product.dto.DescriptionResponse;
 import org.kakaoshare.backend.domain.product.dto.DetailResponse;
-import org.kakaoshare.backend.domain.product.entity.query.SimpleProductDto;
+import org.kakaoshare.backend.domain.product.dto.SimpleProductDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
